+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file. The format 
 ## [v0.1.2] - 2024-04-19
 ### Fixed
 - Make core properties boolable
+- Fix 'umk config write' crash with 'NoneType object is not iterable'
 
 ## [v0.1.1] - 2024-04-19
 ### Fixed