+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 All notable changes to this project will be documented in this file. The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [v0.1.3] - 2024-04-19
+### Change
+- Rename .unimake/remotes.py -> .unimake/remote.py
+
 ## [v0.1.2] - 2024-04-19
 ### Fixed
 - Fix assignment of config entry with type int, bool and float